fix: increase wait time for will message before reconnect
This fixes a flaky test case which asserts will message is sent
if the will delay is shorter than session expiry.

Prior to this fix, the reconnect was done immediately done after
the will message delay.
This might cause a race with the delay timer in the broker.

In this commit, the delay before reconnect is added with 1 extra
second.

This commit also changed the will delay from 5 seconds to 2 to
speed up the test run a bit.